Fix rest of compiler warnings. Updated SDWebImage to add 64bits support.

+++ tvheadend-ios-lib/TVHAdaptersStore.h
@@ -33,7 +33,7 @@
 - (id)initWithTvhServer:(TVHServer*)tvhServer;
 - (void)fetchAdapters;
 
-- (TVHAdapter*)objectAtIndex:(int) row;
+- (TVHAdapter*)objectAtIndex:(NSUInteger) row;
 - (int)count;
 @end
