+++ tvheadend-ios-lib/TVHAdaptersStoreAbstract.h
@@ -23,6 +23,6 @@
 - (id)initWithTvhServer:(TVHServer*)tvhServer;
 - (void)fetchAdapters;
 
-- (TVHAdapter *)objectAtIndex:(int) row;
+- (TVHAdapter *)objectAtIndex:(NSUInteger) row;
 - (int)count;
 @end
