+++ tvheadend-ios-lib/TVHAutoRecStore.h
@@ -29,7 +29,7 @@
 - (id)initWithTvhServer:(TVHServer*)tvhServer;
 - (void)fetchDvrAutoRec;
 
-- (TVHAutoRecItem *)objectAtIndex:(int)row;
+- (TVHAutoRecItem *)objectAtIndex:(NSUInteger)row;
 - (int)count;
 @end
