+++ tvheadend-ios-lib/TVHDvrStoreAbstract.h
@@ -22,7 +22,7 @@
 - (id)initWithTvhServer:(TVHServer*)tvhServer;
 - (void)fetchDvr;
 
-- (TVHDvrItem *)objectAtIndex:(int)row forType:(NSInteger)type;
+- (TVHDvrItem *)objectAtIndex:(NSUInteger)row forType:(NSInteger)type;
 - (int)count:(NSInteger)type;
 - (void)signalDidLoadDvr:(NSInteger)type;
 @end
