+++ tvheadend-ios-lib/TVHNetworkStore32.m
@@ -18,7 +18,7 @@
 {
 }
 
-- (TVHNetwork*)objectAtIndex:(int) row
+- (TVHNetwork*)objectAtIndex:(NSUInteger) row
 {
     return nil;
 }
