+++ tvheadend-ios-lib/TVHStatusInputStore32.h
@@ -18,6 +18,6 @@
 - (id)initWithTvhServer:(TVHServer*)tvhServer;
 - (void)fetchStatusInputs;
 
-- (TVHStatusInput*)objectAtIndex:(int) row;
+- (TVHStatusInput*)objectAtIndex:(NSUInteger) row;
 - (int)count;
 @end
