+++ tvheadend-ios-lib/TVHStatusInputStore40.h
@@ -17,6 +17,6 @@
 - (id)initWithTvhServer:(TVHServer*)tvhServer;
 - (void)fetchStatusInputs;
 
-- (TVHStatusInput*)objectAtIndex:(int) row;
+- (TVHStatusInput*)objectAtIndex:(NSUInteger) row;
 - (int)count;
 @end
